Finalizado o back-end com as funcionalidades necessárias

+++ src/main/java/unoeste/fipp/ativooperante_be/repositories/OrgaosRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import unoeste.fipp.ativooperante_be.entities.Orgaos;
 @Repository
-public interface OrgaosRepository extends JpaRepository<Orgaos, Integer> {
+public interface OrgaosRepository extends JpaRepository<Orgaos, Long> {
 }
